% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parAL-package.R
\docType{data}
\name{orig.weight}
\alias{orig.weight}
\title{Weight matrix orig.weight}
\format{workspace}
\source{
simulation
}
\description{
This object contains a simulated weight matrix that provides as example weight matrix for the examples of the functions.
#'
}
\keyword{matrix}
\keyword{weight}

