% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parAL-package.R
\docType{package}
\name{parAL-package}
\alias{parAL}
\alias{parAL-package}
\title{Goodness-of-fit of the user-defined weight matrix of the linear logistic test model}
\description{
The linear logistic test model (LLTM, Fischer, 1973) is estimated as a restricted Rasch model. For the LLTM, the user has to set up a design matrix. To assess the fit of this user-defined weight matrix, a method is proposed by Baghaei and Hohensinn (submitted) which is analogous to the parallel analysis in factor analysis.
}
\details{
This package allows to conduct this proposed approach for an LLTM.

\tabular{ll}{ Package: \tab parAL\cr Type: \tab Package\cr Version: \tab
0.1\cr Date: \tab 2016-09-26\cr License: \tab GPL-3\cr }
}
\author{
Christine Hohensinn Maintainer: Christine Hohensinn
<christine.hohensinn@univie.ac.at>
}
\references{
Baghaei, P. & Hohensinn, C. (submitted). A Method of Q-Matrix Validation for the Linear Logistic Test Model.

Fischer, G. H. (1973). The linear logistic test model as an instrument in educational research. Acta Psychologica, 37, 359-374.
}
\seealso{
\code{\link{Wrandom}} \code{\link{Wpermut}} \code{\link{Wperfect}}
}
\keyword{IRT}
\keyword{Item}
\keyword{Response}
\keyword{Theory}
\keyword{linear}
\keyword{logistic}
\keyword{model}
\keyword{package}
\keyword{psychometrics}
\keyword{test}

