% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wperfect.R
\name{Wperfect}
\alias{Wperfect}
\title{generate data matrices according to LLTM}
\usage{
Wperfect(datmat, weightmat, repl = 50, keepmat = FALSE)
}
\arguments{
\item{datmat}{data file}

\item{weightmat}{original weight matrix}

\item{repl}{number of weight matrices that are generated}

\item{keepmat}{if TRUE, the generated random data matrices are kept in an three-dimensional array}
}
\value{
\item{convLLTM}{convergence of the estimated LLTMs of the generated data matrices}
\item{convRM}{convergence of the estimated Rasch models of the generated data matrices}
\item{lltmpar}{item parameters reconstructed from the estimated parameters of the LLTM of the generated weight matrices}
\item{raschpar}{estimated Rasch model item parameters for the original data set}
\item{corlr}{correlation of the re-constructed item parameters by the LLTM and the estimated item parameters of the original Rasch model}
\item{sumcorlr}{short summary descriptive statistics of the correlations}
\item{randommat}{array containing the generated data matrices (if argument keepmat is set TRUE)}
}
\description{
generate data matrices according to the original weight matrix
}
\details{
The original data matrix and the original weight matrix are taken and an LLTM is estimated. With the resulting estimated parameters and the original data matrix new random data matrices are generated according to the original weight matrix. For each generated data matrix, the Rasch model and an LLTM is estimated with the original weight matrix. The correlation between the item parameters of the Rasch model and the item parameters reconstructed from the parameters estimated by the LLTM is calculated. With this procedure some kind of an upper benchmark of the goodness-of-fit of the weight matrix is generated.
}
\examples{

#generate ten data sets according to the original weight matrix
 
data(exampledata)
data(exampleweight)

Wperfect(exdat, orig.weight, repl=10)

}
\references{
Baghaei, P. & Hohensinn, C. (submitted). A Method of Q-Matrix Validation for the Linear Logistic Test Model.

Fischer, G. H. (1973). The linear logistic test model as an instrument in educational research. Acta Psychologica, 37, 359-374.
}
\author{
Christine Hohensinn
}
\keyword{data}
\keyword{linear}
\keyword{logistic}
\keyword{matrix,}
\keyword{model}
\keyword{random}
\keyword{test}
