% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wpermut.R
\name{Wpermut}
\alias{Wpermut}
\title{generate permutated weight matrices}
\usage{
Wpermut(datmat, weightmat, repl = 100, keepmat = FALSE)
}
\arguments{
\item{datmat}{data file}

\item{weightmat}{original weight matrix}

\item{repl}{number of weight matrices that are generated}

\item{keepmat}{if TRUE, the generated random weight matrices are kept in an three-dimensional array}
}
\value{
\item{conv}{convergence of the estimated LLTMs of the generated weight matrices}
\item{npermut}{number of generated weight matrices}
\item{nchange}{matrix with the frequencies of the number of rows that were changed in the generated weight matrices}
\item{hchange}{vector with the number of changed rows for each permutation}
\item{raschpar}{estimated Rasch model item parameters for the original data set}
\item{lltmpar}{item parameters reconstructed from the estimated parameters of the LLTM of the generated weight matrices}
\item{corlr}{correlation of the re-constructed item parameters by the LLTM and the estimated item parameters of the original Rasch model}
\item{sumcorlr}{short summary descriptive statistics of the correlations}
\item{randommat}{array containing the generated weight matrices (if argument keepmat is set TRUE)}
}
\description{
generate weight matrices that permutate rows from the original weight matrix
}
\details{
The original weight matrix is taken and new weight matrices are generated by permutating the rows of the original matrix. The permutations are drawn randomly. Duplicated permutations are eliminated - thus the user-defined number of replications can deviate from the number of generated weight matrices. For each generated weight matrix, an LLTM is estimated using the function \code{DRM} of the R package pcIRT. A Rasch model is estimated for the original data set using the function \code{DRM} from pcIRT. The item parameters estimated by the Rasch model are correlated with the item parameters reconstructed from the parameters estimated by the LLTM.
}
\examples{

#generate three permutated weight matrices for example data set
data(exampledata)
data(exampleweight)

Wpermut(exdat, orig.weight, repl=3)

}
\references{
Baghaei, P. & Hohensinn, C. (submitted). A Method of Q-Matrix Validation for the Linear Logistic Test Model.

Fischer, G. H. (1973). The linear logistic test model as an instrument in educational research. Acta Psychologica, 37, 359-374.
}
\author{
Christine Hohensinn
}
\keyword{linear}
\keyword{logistic}
\keyword{matrix,}
\keyword{model}
\keyword{random}
\keyword{test}
\keyword{weight}
