% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrandom.R
\name{Wrandom}
\alias{Wrandom}
\title{generate random weight matrices}
\usage{
Wrandom(datmat, weightmat, repl = 100, prop01 = c(0.5, 0.5),
  keepmat = FALSE)
}
\arguments{
\item{datmat}{data file}

\item{weightmat}{original weight matrix}

\item{repl}{number of weight matrices that are generated}

\item{prop01}{vector providing the proportions of Zeros and Ones of the weight matrix to be generated}

\item{keepmat}{if TRUE, the generated random weight matrices are kept in an three-dimensional array}
}
\value{
\item{conv}{convergence of the estimated LLTMs of the generated weight matrices}
\item{raschpar}{estimated Rasch model item parameters for the original data set}
\item{lltmpar}{item parameters reconstructed from the estimated parameters of the LLTM of the generated weight matrices}
\item{corlr}{correlation of the re-constructed item parameters by the LLTM and the estimated item parameters of the original Rasch model}
\item{sumcorlr}{short summary descriptive statistics of the correlations }
\item{randommat}{array containing the generated weight matrices (if argument keepmat is set TRUE)}
}
\description{
generate random weight matrices according to the dimensions of the original weight matrix
}
\details{
According to an original data file with a defined weight matrix to estimate a linear logistic test model (LLTM), random weight matrices with the same dimensions of the original one are generated. For each generated weight matrix, an LLTM is estimated using the function \code{DRM} of the R package pcIRT. A Rasch model is estimated for the original data set using the function \code{DRM} from pcIRT. The item parameters estimated by the Rasch model are correlated with the item parameters reconstructed from the parameters estimated by the LLTM.
}
\examples{

#generate three random weight matrices for example data set
data(exampledata)
data(exampleweight)

Wrandom(exdat, orig.weight, repl=3)

}
\references{
Baghaei, P. & Hohensinn, C. (submitted). A Method of Q-Matrix Validation for the Linear Logistic Test Model.

Fischer, G. H. (1973). The linear logistic test model as an instrument in educational research. Acta Psychologica, 37, 359-374.
}
\author{
Christine Hohensinn
}
\keyword{linear}
\keyword{logistic}
\keyword{matrix,}
\keyword{model}
\keyword{random}
\keyword{test}
\keyword{weight}
